Fix file paths when globbing default
This fixes the order of files globbed by default so that they are
consistent with the order loaded from the config file. The root
path is stripped off of the globbed files. The paths of the ordered
files are now always relative to the root path of the proze project.